shuffle code around
